Corrected diameter mesurement of droplets
Using a new algorithm. details in the code.
basically labview gives the centre of the droplets. I filter the
duplicates and go in in circles from the centre, capturing the highest
value. I look for a positive derivative and catch the value where in
returns to white values.
Changed default diameter filter sizes.